requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
openai>=1.0.0 